[Back](users tests) add tests on users service and controller (#98)
* add unique mail constraint

* add error handling on unique email constraint violation

* tests of users service

* fix according to the tests done

* fix creation payload after rebase

* fix problematic migrations

* add tests about asso of an user

* tests on users controller

* fix users controller according to tests

* fix code smells

+++ libs/backend/feature/user/src/lib/users.service.ts
@@ -28,14 +28,33 @@ export class UsersService {
   }
 
   public async findOne(id: number): Promise<UserDto> {
-    return this.userRepository.findOne(id);
+    const user = await this.userRepository.findOne(id);
+    if (!user) {
+      throw new Error('User not found');
+    }
+    return user;
   }
 
   public async update(id: number, updateBaseDto: UpdateUserDto): Promise<UpdateResult> {
-    return this.userRepository.update(id, updateBaseDto);
+    const user = this.userRepository.findOne(id);
+    if (!user) {
+      throw new Error('User not found');
+    }
+
+    try {
+      return await this.userRepository.update(id, updateBaseDto);
+    } catch (error) {
+      if (error?.code === PostgresError.UNIQUE_VIOLATION) {
+        throw new Error('Email already used');
+      }
+    }
   }
 
   public async delete(id: number): Promise<UpdateResult> {
+    const user = this.userRepository.findOne(id);
+    if (!user) {
+      throw new Error('User not found');
+    }
     return this.userRepository.delete(id);
   }
